+++ lib/graphql_rails/controller/log_controller_action.rb
@@ -51,8 +51,7 @@ module GraphqlRails
             params
           else
             filter_options = Rails.configuration.filter_parameters
-            parametter_filter = ActionDispatch::Http::ParameterFilter.new(filter_options)
-            parametter_filter.filter(params)
+            parameter_filter_class.new(filter_options).filter(params)
           end
       end
 
@@ -60,6 +59,12 @@ module GraphqlRails
         return [] if !defined?(Rails) || Rails.application.nil?
 
         Rails.application.config.filter_parameters || []
+      end
+
+      def parameter_filter_class
+        return ActiveSupport::ParameterFilter if Object.const_defined?('ActiveSupport::ParameterFilter')
+
+        ActionDispatch::Http::ParameterFilter
       end
     end
   end